+++ backend-flask/lib/rollbar.py
@@ -1,10 +1,9 @@
-from flask import current_app as app
 from flask import got_request_exception
 import os
 import rollbar
 import rollbar.contrib.flask
 
-def init_rollbar():
+def init_rollbar(app):
     rollbar_access_token = os.getenv('ROLLBAR_ACCESS_TOKEN')
     rollbar.init(
         # access token
